Ability System Updates
Created base ability class to allow the creation of abilities as scriptable objects.
Dash ability template created. Logic for dash needs to be implemented. 
AbilityHolder class is created to allow the player to hold one or more ability objects. 
Made local changes to the Ability UI. No prefabs for ability UI is created.